requirements.txt
attrs==20.2.0
aws-cdk.assets==1.71.0
aws-cdk.aws-apigatewayv2==1.71.0
aws-cdk.aws-applicationautoscaling==1.71.0
aws-cdk.aws-autoscaling-common==1.71.0
aws-cdk.aws-certificatemanager==1.71.0
aws-cdk.aws-cloudformation==1.71.0
aws-cdk.aws-cloudwatch==1.71.0
aws-cdk.aws-codeguruprofiler==1.71.0
aws-cdk.aws-dynamodb==1.71.0
aws-cdk.aws-ec2==1.71.0
aws-cdk.aws-efs==1.71.0
aws-cdk.aws-events==1.71.0
aws-cdk.aws-iam==1.71.0
aws-cdk.aws-kms==1.71.0
aws-cdk.aws-lambda==1.71.0
aws-cdk.aws-lambda-python==1.71.0
aws-cdk.aws-logs==1.71.0
aws-cdk.aws-route53==1.71.0
aws-cdk.aws-s3==1.71.0
aws-cdk.aws-s3-assets==1.71.0
aws-cdk.aws-sns==1.71.0
aws-cdk.aws-sqs==1.71.0
aws-cdk.aws-ssm==1.71.0
aws-cdk.cloud-assembly-schema==1.71.0
aws-cdk.core==1.71.0
aws-cdk.custom-resources==1.71.0
aws-cdk.cx-api==1.71.0
aws-cdk.region-info==1.71.0
boto3==1.16.9
botocore==1.19.9
cattrs==1.1.1
constructs==3.2.8
jmespath==0.10.0
jsii==1.14.0
publication==0.0.3
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.25.11
